refactor: ViewController 추가 및 사용되지 않는 클래스 제거
- AdminController 제거
- 기존 Controller > RestController로 변경
