+++ src/main/java/roomescape/admin/controller/ReservationTimeController.java
@@ -1,7 +1,6 @@
 package roomescape.admin.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import roomescape.admin.dto.ReadReservationTimeResponse;
 import roomescape.admin.dto.SaveReservationTimeRequest;
@@ -9,19 +8,12 @@ import roomescape.admin.service.ReservationTimeService;
 
 import java.util.List;
 
-@Controller
+@RestController
 public class ReservationTimeController {
-    private static final String BASE_PATH = "/admin";
-
     private final ReservationTimeService reservationTimeService;
 
     public ReservationTimeController(ReservationTimeService reservationTimeService){
         this.reservationTimeService = reservationTimeService;
-    }
-
-    @GetMapping("/time")
-    public String time(){
-        return BASE_PATH+"/time.html";
     }
 
     @GetMapping("/times")
