+++ src/main/java/roomescape/admin/controller/ViewController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class AdminController {
+public class ViewController {
 
     private static final String BASE_PATH = "/admin";
 
@@ -13,4 +13,13 @@ public class AdminController {
         return BASE_PATH+"/index.html";
     }
 
+    @GetMapping("/reservation")
+    public String reservation(){
+        return BASE_PATH+"/reservation.html";
+    }
+
+    @GetMapping("/time")
+    public String time(){
+        return BASE_PATH+"/time.html";
+    }
 }
